requirements.txt
azure-identity
azure-mgmt-compute 
azure-mgmt-network 
azure-mgmt-resource
azure-mgmt-storage
azure-storage-blob 
pandas
